LayoutTests/platform/ios-simulator/ios/fast/coordinates/touch-event-expected.txt

This tests touch event coordinates unscaled.
On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
touchstart
PASS eventOrTouch.pageX is 125
PASS eventOrTouch.pageY is 175
PASS eventOrTouch.clientX is 125
PASS eventOrTouch.clientY is 175
touchmove
PASS eventOrTouch.pageX is 175
PASS eventOrTouch.pageY is 125
PASS eventOrTouch.clientX is 175
PASS eventOrTouch.clientY is 125
touchend - no touch to check
PASS successfullyParsed is true

TEST COMPLETE
